requirements.txt
pandas
sqlalchemy
pymysql
matplotlib
python-dotenv